Adapt SetupTransportPinViewModel tests
